refactor(agent): align protocol module layout
