+++ docs/backends/extend.zh.md
@@ -63,7 +63,7 @@ Backend Adapter SDK（`qwen-audio-agent/backend-adapter-sdk`）直接实现
 
 ### 原料二：Agent Driver（必填）
 
-`server/src/agent/backends/<id>.mjs`，外加 `registry.mjs` 里的一行
+`server/src/agent/acp/drivers/<id>.mjs`，外加 `registry.mjs` 里的一行
 导入。Driver 声明能力契约并构建运行时档案：
 
 ```js
@@ -92,7 +92,7 @@ export const myBackendDriver = {
 ```
 
 八个能力标志都是必填布尔值——注册中心会校验契约，
-`backend-driver-registry.test.mjs` 会断言每个对外宣称的后台都有完整的
+`acp-driver-registry.test.mjs` 会断言每个对外宣称的后台都有完整的
 driver。注册了一半的后台会在启动时响亮地失败，而不是在运行时静默出错。
 
 ### 原料三：Runtime Driver（按需）
@@ -109,7 +109,7 @@ runtime driver 的后台（外部服务支持、自定义拉起规则）。
 2. 一等公民后台：在 `shared/backend-catalog.mjs` 加目录条目，加 agent
    driver 并注册，按需加 runtime driver 处理自定义进程归属。
 3. 跑契约测试——它们是门禁：ACP driver 跑
-   `node --test server/test/backend-driver-registry.test.mjs`，自定义
+   `node --test server/test/acp-driver-registry.test.mjs`，自定义
    适配器跑 `node --test server/test/backend-adapter-sdk.test.mjs`。
 
 ## 继续阅读
