+++ server/src/agent/acp/builtin-mcp.mjs
