+++ server/src/agent/acp/drivers/opencode.mjs
