+++ server/src/agent/acp/request-signal.mjs
@@ -1,0 +1,5 @@
+export function requestSignal(signal, timeoutMs) {
+  if (!timeoutMs) return signal
+  const timeout = AbortSignal.timeout(timeoutMs)
+  return signal ? AbortSignal.any([signal, timeout]) : timeout
+}
