+++ server/src/agent/agent-client.mjs
@@ -1,6 +1,6 @@
 import { config } from '../core/config.mjs'
 import { assertBackendPort } from '../backend/backend-port.mjs'
-import { AgentError } from './backend-adapter.mjs'
+import { AgentError } from './agent-error.mjs'
 import { createAcpBackendAdapter } from './acp/backend-factory.mjs'
 
 export { AgentError }
