+++ server/src/agent/agent-error.mjs
@@ -7,9 +7,3 @@ export class AgentError extends Error {
     this.protocol = protocol
   }
 }
-
-export function requestSignal(signal, timeoutMs) {
-  if (!timeoutMs) return signal
-  const timeout = AbortSignal.timeout(timeoutMs)
-  return signal ? AbortSignal.any([signal, timeout]) : timeout
-}
