+++ server/src/app/gateway-application.mjs
@@ -3,7 +3,7 @@ import { createServer } from 'http'
 import { randomUUID } from 'node:crypto'
 import { resolve } from 'path'
 import { agent as defaultAgent } from '../agent/agent-client.mjs'
-import { BackendAvailability } from '../agent/backend-availability.mjs'
+import { BackendAvailability } from '../backend/availability.mjs'
 import { BackendWorkRuntime } from '../backend/backend-work-runtime.mjs'
 import { config as defaultConfig } from '../core/config.mjs'
 import { logger as defaultLogger, runWithLogContext } from '../core/logger.mjs'
