+++ server/test/acp-process-client.test.mjs
@@ -3,7 +3,7 @@ import { EventEmitter } from 'node:events'
 import { PassThrough } from 'node:stream'
 import test from 'node:test'
 import { AcpProcessClient } from '../src/agent/acp/process-client.mjs'
-import { openClawBackendDriver } from '../src/agent/backends/openclaw.mjs'
+import { openClawBackendDriver } from '../src/agent/acp/drivers/openclaw.mjs'
 
 test('keeps session object identity stable across re-registration', () => {
   const client = new AcpProcessClient({
